Make sure the functions end up in the package
index.js is included to ensure the individual function files end up in
the package.
That way oter packages might use a single function like so:

```
import head from '@grrr/util/functions/head';
```

That's also the reason I renamed the folder from `src` to `functions`,
because it is now an exposed folder.